requirements.txt
flet
beautifulSoup4
requests
Pillow